tests/wpt/tests/css/css-align/parsing/gap-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Box Alignment Level 3: parsing gap with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-gap">
<meta name="assert" content="gap supports only the grammar '<row-gap> <column-gap>?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("gap", "auto");
test_invalid_value("gap", "-10px");

test_invalid_value("gap", "10px 20% 30px");
test_invalid_value("gap", "normal 10px normal");
</script>
</body>
</html>
